+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
  - Add `onionperf measure --drop-guards` parameter to use and drop
    guards after a given number of hours. Implements #33399.
+ - Remove the `onionperf measure --oneshot` switch and replace it with
+   new switches `--tgen-pause-initial`, `--tgen-pause-between`,
+   `--tgen-transfer-size`, and `--tgen-num-transfers ` to further
+   configure the generated TGen model.
 
 # Changes in version 0.6 - 2020-08-08